requirements.txt
fastapi
uvicorn
python-dotenv
motor
email-validator
celery 
redis 
sib-api-v3-sdk 